requirements.txt
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pytesseract==0.3.6
requests==2.25.0
selenium==3.141.0
tesseract==0.1.3
urllib3==1.26.2
webdriver-manager==3.2.2
